Export single IVSUIContextFactory implementation in GH.TF.14
This allows us to consume IVSUIContextFactory as a MEF service and simplify unit tests.
