+++ src/GitHub.TeamFoundation.14/GitHub.TeamFoundation.14.csproj
@@ -134,6 +134,7 @@
     <Compile Include="Services\VSGitExt.cs" />
     <Compile Include="RegistryHelper.cs" />
     <Compile Include="Services\VSGitServices.cs" />
+    <Compile Include="Services\VSUIContextFactory.cs" />
     <Compile Include="Settings.cs" />
     <Compile Include="Base\EnsureLoggedInSection.cs" />
     <Compile Include="Base\TeamExplorerInvitationBase.cs" />
